+++ dadapy/__init__.py
@@ -2,4 +2,4 @@ from .clustering import *
 from .data import *
 from .density_estimation import *
 from .id_estimation import *
-from .utils_.utils import *
+from ._utils.utils import *
